+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 - Allow passing Python mapping as arguments of type `Iterable[IKeyValuePair[K, V]]`.
+- Added ``get()`` and ``wait()`` methods to async operations to call them synchronously.
 
 ## [v3.1.0] - 2025-04-11